Fixing a modulo issue with negative values

See merge request f3d/f3d!108

Former-commit-id: 7c13b3bc98fd86db363515ab13f26a8905e4e142

+++ src/F3DLoader.h
@@ -63,7 +63,7 @@ protected:
   static vtkSmartPointer<vtkImporter> GetImporter(const F3DOptions& options, const std::string& file);
 
   std::vector<std::string> FilesList;
-  size_t CurrentFileIndex = 0;
+  int CurrentFileIndex = 0;
   F3DOptionsParser Parser;
   vtkSmartPointer<vtkF3DRenderer> Renderer;
